+++ account/src/vonage_account/__init__.py
@@ -1,3 +1,12 @@
 from .account import Account
+from .errors import InvalidSecretError
+from .responses import Balance, SettingsResponse, TopUpResponse, VonageApiSecret
 
-__all__ = ['Account']
+__all__ = [
+    'Account',
+    'Balance',
+    'InvalidSecretError',
+    'SettingsResponse',
+    'TopUpResponse',
+    'VonageApiSecret',
+]