+++ account/src/vonage_account/errors.py
@@ -1,5 +1,5 @@
 from vonage_utils.errors import VonageError
 
 
-class AccountError(VonageError):
-    """Indicates an error with the Account package."""
+class InvalidSecretError(VonageError):
+    """Indicates that the secret provided was invalid."""