+++ application/CHANGES.md
@@ -1,2 +1,5 @@
+# 1.0.1
+- Update project metadata
+
 # 1.0.0
 - Initial upload