+++ vonage_network/CHANGES.md
@@ -1,2 +1,5 @@
+# 0.1.1b0
+- Update project metadata
+
 # 0.1.0b0
 - Initial beta release